+++ src/components/FallbackScreens/LoadingScreen.tsx
@@ -1,4 +1,5 @@
 import { Spinner } from "@fluentui/react-components";
+
 export const LoadingScreen = () => {
   return <Spinner />;
 };
