+++ src/components/FallbackScreens/index.ts
@@ -3,3 +3,4 @@ export * from "./LoadingScreen";
 export * from "./NoDataScreen";
 export * from "./FallbackScreen";
 export * from "./FallbackScreen.types";
+export * from "./Unathorized";
