+++ src/components/Toolbar/Toolbar.tsx
@@ -3,30 +3,17 @@ import {
   Toolbar as ToolbarComponent,
   Link as FluentLink,
   Button,
-  Avatar,
-  Tooltip,
-  Badge,
-  Menu,
-  MenuItem,
-  MenuTrigger,
-  MenuPopover,
-  MenuList,
-  MenuDivider,
+  mergeClasses,
   tokens,
 } from "@fluentui/react-components";
-import {
-  BookOpenRegular,
-  PersonCircleRegular,
-  SettingsRegular,
-  Add24Regular,
-  CollectionsRegular,
-} from "@fluentui/react-icons";
-import { LargeTitle } from "../";
+import { Add24Regular } from "@fluentui/react-icons";
 import { makeStyles, shorthands } from "@griffel/react";
 import { SearchBar } from "./SearchBar";
 import { useRouter } from "next/router";
 import Link from "next/link";
-import { useSession, signIn, signOut } from "next-auth/react";
+import { useSession, signIn } from "next-auth/react";
+import { UserProfile } from "./UserProfile";
+import { Logo } from "./Logo";
 
 const useToolbarStyles = makeStyles({
   root: {
@@ -101,74 +88,12 @@ const useToolbarStyles = makeStyles({
     },
   },
   activeLink: {
-    color: tokens.colorBrandForeground1,
     backgroundColor: tokens.colorNeutralBackground3,
-    fontWeight: "600",
-    textDecoration: "none",
-    ":after": {
-      content: "''",
-      position: "absolute",
-      bottom: "-4px",
-      left: "30%",
-      right: "30%",
-      height: "3px",
-      backgroundColor: tokens.colorBrandBackground,
-      borderRadius: "1.5px",
-      boxShadow: "0 1px 2px rgba(105, 67, 209, 0.3)",
-    },
-  },
-  linkIcon: {
-    display: "flex",
-    fontSize: "16px",
-  },
-
-  logo: {
-    display: "flex",
-    alignItems: "center",
-    gap: "10px",
-    fontWeight: "700",
-    fontSize: "20px",
-    color: tokens.colorBrandBackground,
-    textDecoration: "none",
-    transition: "transform 0.2s ease",
-    ":hover": {
-      transform: "scale(1.02)",
-    },
-  },
-  logoIcon: {
-    display: "flex",
-    alignItems: "center",
-    justifyContent: "center",
-    width: "34px",
-    height: "34px",
-    backgroundColor: tokens.colorBrandBackground,
-    color: "white",
-    borderRadius: "10px",
-    boxShadow: "0 2px 5px rgba(146, 114, 230, 0.3)",
   },
   navLinks: {
     display: "flex",
     gap: "8px",
     paddingLeft: "32px",
-  },
-  iconButton: {
-    ...shorthands.borderRadius("8px"),
-    width: "40px",
-    height: "40px",
-    backgroundColor: "transparent",
-    color: tokens.colorNeutralForeground2,
-    ":hover": {
-      backgroundColor: tokens.colorNeutralBackground2Hover,
-      color: tokens.colorBrandForeground1,
-    },
-    transition: "all 0.2s ease",
-  },
-  avatarButton: {
-    cursor: "pointer",
-    ":hover": {
-      opacity: 0.8,
-    },
-    transition: "transform 0.2s ease",
   },
   toolbarButton: {
     display: "flex",
@@ -190,27 +115,8 @@ const useToolbarStyles = makeStyles({
       boxShadow: "0 4px 12px rgba(105, 67, 209, 0.25)", // More pronounced shadow
     },
   },
-
-  logoText: {
-    background: tokens.colorBrandBackground,
-    backgroundClip: "text",
-    WebkitBackgroundClip: "text",
-    WebkitTextFillColor: "transparent",
-  },
-
   nextLink: {
     textDecoration: "none",
-  },
-
-  menuItem: {
-    display: "flex",
-    alignItems: "center",
-    gap: "8px",
-  },
-
-  menuIcon: {
-    display: "flex",
-    fontSize: "16px",
   },
 });
 
@@ -223,74 +129,52 @@ export const Toolbar = () => {
   return (
     <ToolbarComponent className={styles.root}>
       <div className={styles.leftSection}>
-        <Link href="/" passHref className={styles.nextLink}>
-          <FluentLink
-            className={styles.logo}
-            style={{ textDecoration: "none" }}
-          >
-            <span className={styles.logoIcon}>
-              <BookOpenRegular fontSize={18} />
-            </span>
-            <span className={styles.logoText}>Book Cook</span>
-          </FluentLink>
-        </Link>
-
-        <div className={styles.navLinks}>
-          <Link href="/recipes" className={styles.nextLink} passHref>
-            <FluentLink
-              className={`${styles.linkStyles} ${path.includes("/recipes") ? styles.activeLink : ""}`}
-            >
-              Recipes
-            </FluentLink>
-          </Link>
-          {/* <Link href="/collections" className={styles.nextLink} passHref>
-            <FluentLink
-              className={`${styles.linkStyles} ${path.includes("/collections") ? styles.activeLink : ""}`}
-            >
-              Collections
-            </FluentLink>
-          </Link>
-          <Link href="/explore" className={styles.nextLink} passHref>
-            <FluentLink
-              className={`${styles.linkStyles} ${path.includes("/explore") ? styles.activeLink : ""}`}
-            >
-              Explore
-            </FluentLink>
-          </Link> */}
-        </div>
+        <Logo />
+        {session?.user && (
+          <div className={styles.navLinks}>
+            <Link href="/recipes" className={styles.nextLink} passHref>
+              <FluentLink
+                className={mergeClasses(
+                  styles.linkStyles,
+                  path.includes("/recipes") && styles.activeLink
+                )}
+              >
+                Recipes
+              </FluentLink>
+            </Link>
+            <Link href="/collections" className={styles.nextLink} passHref>
+              <FluentLink
+                className={`${styles.linkStyles} ${path.includes("/collections") ? styles.activeLink : ""}`}
+              >
+                Collections
+              </FluentLink>
+            </Link>
+            <Link href="/explore" className={styles.nextLink} passHref>
+              <FluentLink
+                className={`${styles.linkStyles} ${path.includes("/explore") ? styles.activeLink : ""}`}
+              >
+                Explore
+              </FluentLink>
+            </Link>
+          </div>
+        )}
       </div>
       <div className={styles.rightSection}>
-        <Button
-          appearance="primary"
-          icon={<Add24Regular />}
-          onClick={() => router.push("/newRecipe")}
-          className={styles.toolbarButton}
-        >
-          New Recipe
-        </Button>
-        <SearchBar />
+        {session?.user && (
+          <>
+            <Button
+              appearance="primary"
+              icon={<Add24Regular />}
+              onClick={() => router.push("/newRecipe")}
+              className={styles.toolbarButton}
+            >
+              New Recipe
+            </Button>
+            <SearchBar />
+          </>
+        )}
         {session?.user ? (
-          <Menu>
-            <MenuTrigger>
-              <Tooltip content="Profile" relationship="label">
-                <Avatar
-                  name={session.user.name || session.user.email || "Unknown"}
-                  size={36}
-                  color="brand"
-                  style={{ cursor: "pointer" }}
-                />
-              </Tooltip>
-            </MenuTrigger>
-            <MenuPopover>
-              <MenuList>
-                <MenuItem onClick={() => router.push("/settings")}>
-                  Settings
-                </MenuItem>
-                <MenuDivider />
-                <MenuItem onClick={() => signOut()}>Sign Out</MenuItem>
-              </MenuList>
-            </MenuPopover>
-          </Menu>
+          <UserProfile />
         ) : (
           <Button appearance="primary" onClick={() => signIn("google")}>
             Sign In
