+++ src/components/Toolbar/UserProfile/index.ts
@@ -1,0 +1,1 @@
+export * from "./UserProfile";
