+++ src/components/index.ts
@@ -5,3 +5,5 @@ export * from "./MarkdownParser";
 export * from "./FallbackScreens";
 export * from "./RecipeCard";
 export * from "./TagPicker";
+export * from "./LandingPage";
+export * from "./HomePage";
