+++ src/pages/settings.tsx
@@ -1,0 +1,7 @@
+import * as React from "react";
+
+import { useSession } from "next-auth/react";
+
+export default function Settings() {
+  return <div>Settings</div>;
+}
